Move handler to composable
to keep the logic related to lessons
loading in one place
